Add sensor configuration
Add sensor_configs to configuration. This is built similarly as the
light_config, using JSONB in the database to simplify storage and
iteration, but stores an array instead of an object.

Currently only the AM2320 sensor configuration is defined. This entire
approach needs to be reviewed to detach feature targets and
implementations details (using a hardward/feature/target approach?).

Also adds changes to Kubernetes file that were missing.